requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-allauth==0.41.0
django-crispy-forms==1.12.0
djangorestframework==3.12.4
gunicorn==20.1.0
oauthlib==3.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.60.0
